requirements.txt
# Core dependencies
pyyaml>=6.0
httpx>=0.24
feedparser>=6.0
jinja2>=3.1
python-dateutil>=2.8

# AI
anthropic>=0.18

# Optional (for card rendering)
playwright>=1.40

# Development
pytest>=7.0
black>=23.0
